feat: introduce recipe card component and initial favorite recipe screen

+++ Rango/Rango/Model/Recipe.swift
@@ -16,7 +16,7 @@ public struct Recipe: Decodable {
     let bio: String
     let dificulty: Dificult
     let time: Int
-    let rating: Int
+    let rating: Double
     let image: String
     let ingredients: [String]
     let steps: [String]
